Added casts

+++ src/main/telemetry/crsf.c
@@ -294,8 +294,8 @@ static void crsfFrameAttitude(sbuf_t *dst)
      sbufWriteU8(dst, CRSF_FRAME_ATTITUDE_PAYLOAD_SIZE + CRSF_FRAME_LENGTH_TYPE_CRC);
      crsfSerialize8(dst, CRSF_FRAMETYPE_ATTITUDE);
      crsfSerialize16(dst, DECIDEGREES_TO_RADIANS10000(attitude.values.pitch));
-     crsfSerialize16(dst, flags & 0xffff);  //Used to be roll sending bottom half
-     crsfSerialize16(dst, flags >> 16 );    //Used to be yaw sending top half
+     crsfSerialize16(dst, (int16_t)flags & 0xffff);  //Used to be roll sending bottom half
+     crsfSerialize16(dst, (int16_t)flags >> 16 );    //Used to be yaw sending top half
 }
 
 /*
